// SDKs/ptv.vic.gov.au/v3/typescript/src/sdk/models/shared/v3stopstaffing.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class V3StopStaffing extends SpeakeasyBase {
  @Metadata({ data: "json, name=fri_am_from" })
  friAmFrom?: string;

  @Metadata({ data: "json, name=fri_am_to" })
  friAmTo?: string;

  @Metadata({ data: "json, name=fri_pm_from" })
  friPmFrom?: string;

  @Metadata({ data: "json, name=fri_pm_to" })
  friPmTo?: string;

  @Metadata({ data: "json, name=mon_am_from" })
  monAmFrom?: string;

  @Metadata({ data: "json, name=mon_am_to" })
  monAmTo?: string;

  @Metadata({ data: "json, name=mon_pm_from" })
  monPmFrom?: string;

  @Metadata({ data: "json, name=mon_pm_to" })
  monPmTo?: string;

  @Metadata({ data: "json, name=ph_additional_text" })
  phAdditionalText?: string;

  @Metadata({ data: "json, name=ph_from" })
  phFrom?: string;

  @Metadata({ data: "json, name=ph_to" })
  phTo?: string;

  @Metadata({ data: "json, name=sat_am_from" })
  satAmFrom?: string;

  @Metadata({ data: "json, name=sat_am_to" })
  satAmTo?: string;

  @Metadata({ data: "json, name=sat_pm_from" })
  satPmFrom?: string;

  @Metadata({ data: "json, name=sat_pm_to" })
  satPmTo?: string;

  @Metadata({ data: "json, name=sun_am_from" })
  sunAmFrom?: string;

  @Metadata({ data: "json, name=sun_am_to" })
  sunAmTo?: string;

  @Metadata({ data: "json, name=sun_pm_from" })
  sunPmFrom?: string;

  @Metadata({ data: "json, name=sun_pm_to" })
  sunPmTo?: string;

  @Metadata({ data: "json, name=thu_am_from" })
  thuAmFrom?: string;

  @Metadata({ data: "json, name=thu_am_to" })
  thuAmTo?: string;

  @Metadata({ data: "json, name=thu_pm_from" })
  thuPmFrom?: string;

  @Metadata({ data: "json, name=thu_pm_to" })
  thuPmTo?: string;

  @Metadata({ data: "json, name=tue_am_from" })
  tueAmFrom?: string;

  @Metadata({ data: "json, name=tue_am_to" })
  tueAmTo?: string;

  @Metadata({ data: "json, name=tue_pm_from" })
  tuePmFrom?: string;

  @Metadata({ data: "json, name=tue_pm_to" })
  tuePmTo?: string;

  @Metadata({ data: "json, name=wed_am_from" })
  wedAmFrom?: string;

  @Metadata({ data: "json, name=wed_am_to" })
  wedAmTo?: string;

  @Metadata({ data: "json, name=wed_pm_To" })
  wedPmTo?: string;

  @Metadata({ data: "json, name=wed_pm_from" })
  wedPmFrom?: string;
}